Implement JobLifeCycle Object, use IDispoble for Task, Job, Stream.
This is a big aggregated change on the Prajna Project. This pull request is not fully ready to check-in yet. But I will use this pull request to initiate a code-review request, and
disucss the major changes incorporated in this version.

The main functions implemented are with regard to distributed exception handling and distributed resource management. Our goal will be: 1) Job may throw exception at remote, and all exception
will be serialized and shown at local. 2) We should properly implement handling of IDisposable object throughout our project.

1) JobLifeCycle object.

Each Prajna action (read, fold, etc.) is now associated with a GUID, which is further associated with a JobLifeCycle object. The JobLifeCycle governs the life cycle of a Prajna job,
and is responsible to release any resource used by Job (at App, Daemon, or at Container).

You can register callback (event) on the JobLifeCycle as follows.

OnDispose/OnDisposeFS: function to be called when the job is to be disposed. (only once right before job is disposed).
OnException/OnExceptionFS: function to be called when an exception is encountered. (may be called multiple times if multiple exception is encountered)
OnCancellation/OnCancellationFS: function to be called when the job is cancelled. (only once per job)

JobLifeCycle implements IDisposable interface.

2) SingleJobActionApp/SingleJobActionContainer/SingleJobActionDaemon

JobLifeCycle should not usually be used. Instead, function that plan to do something on the job should attempt to grab a SingleJobActionApp/SingleJobActionContainer/SingleJobActionDaemon object.
SingleJobAction* object also implements IDisposable interface. If a SingleJobAction* object has been successfully grabbed, the subsequent operation can safely use the resource of the job
without fearing that the resource is disposed. The operating logic is that when a SingleJobAction* object will try to increment the reference count of JobLifeCycle when it is created, and
check whether the job is still properly running. If not, try to grab SingleJobAction* object will return a null pointer, and the subsequent operations should not be executed.
When it is disposed, it will attempt to decrement the reference count. When Job is cancelled and the JobLifeCycle object is disposed, it will also decrement the reference count, and mark
the Job as being disposed. Only when the reference count goes to negative, the process will kick in to dispose the job resources.

3) ThrowExceptionAtCallback/EncounterExceptionAtCallback
   /ThrowExceptionAtContainer/EncounterExceptionAtContainer:

These functions are being used to catch exception at remote, aggregate exception, and throw at the local. Ultimately, we should move to a stage that most exceptions at remote (including our
own programming error) should be manifested at local.

4) Job/Task: disposable.

Both Job and Task become disposable object. Thus, they can properly free their resource if they fail.

5) MemStream/StreamBase<byte>, etc..

I start to use the disposable feature of MemStream/StreamBase<byte> to govern stream memory allocation and resource release. My thought is as follows.
a) Application that allocates the stream should be responsible to free its resource. That is, we should be able to
	use ms = new MemStream() ... for most of our jobs.
b) If other function needs to retain a copy of the stream, it should use .Replicate() to get a duplicated copy of the stream.
c) We should try to get rid of .AddRef() and .DecRef() management of the memory stream.