Distributed Function: Registration.
Register a DistributedFunction now return a RegisteredDistributedFunction class. This class provides two interface:
1. Implement IDisposable interface, when disposed, the function is unregistered and no longer respond to the distributed Function call.

2. GetSchemas, which return a 4-tuple that identified the provider, domain, schemaIn and schemaOut of the function registered.

The GetSchemas() can be used to retrieve the signature of the registered DistributedFunction, and allow the function to be accessed agnostically
(e.g., from Prajna Hub, from other language, such as Java, etc.)

JobDependencies.GetInformation( id ) can be used to retrieve the information of the serializer/deserializer. .

Add UnitTest to validate the changes.